pyproject.toml
[tool.poetry]
name = "equipment-backend"
version = "0.1.0"
description = ""
authors = ["fan9704 <cxz123499@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
setuptools = "^69.2.0"
requests = "^2.31.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
openai = "^1.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"